requirements.txt
python-multipart>=0.0.6
fastapi>=0.115.4
uvicorn==0.24.0.post1
python-dotenv==1.0.0
urlextract==1.8.0
pandas==2.2.0
typing-inspect

mammoth==1.7.1
beautifulsoup4==4.12.3
pyarrow==16.0.0
pydantic>=2.7.1
redis==5.2.0

# Build Nov 7.1
SemanticDocumentParser @ git+https://github.com/YorkUITInnovation/SemanticDocumentParser.git

# Don't freak out, the token is read-only on CriadexSDK
# Due for renewal: November 20th, 2024
# v1.2.0
criadexsdk @ git+https://github.com/YorkUITInnovation/CriadexSDK.git